LPS-113884 Move all fragment configuration fields to a single directory
